Generalize Page model
- Add AbstractBasePage for attributes and methods shared with other page types
- Keep tree-related attributes and methods in the page model

(Prerequisite for #357)
