+++ src/cms/models/pages/__init__.py
@@ -1,5 +1,7 @@
 """
-This package contains both page-related data models:
-:class:`~cms.models.pages.page.Page` and
-:class:`~cms.models.pages.page_translation.PageTranslation`
+This package contains page-related data models:
+
+* :class:`~cms.models.pages.abstract_base_page.AbstractBasePage`
+* :class:`~cms.models.pages.page.Page` and
+  :class:`~cms.models.pages.page_translation.PageTranslation`
 """
